+++ OndselSolver/ASMTSpatialContainer.h
@@ -80,15 +80,15 @@ namespace MbD {
         void readOmega3D(std::vector<std::string>& lines);
         void setVelocity3D(double a, double b, double c);
         void setOmega3D(double a, double b, double c);
-        void storeOnLevel(std::ofstream& os, int level) override;
-        void storeOnLevelVelocity(std::ofstream& os, int level);
-        void storeOnLevelOmega(std::ofstream& os, int level);
-        void storeOnLevelRefPoints(std::ofstream& os, int level);
-        void storeOnLevelRefCurves(std::ofstream& os, int level);
-        void storeOnLevelRefSurfaces(std::ofstream& os, int level);
+        void storeOnLevel(std::ofstream& os, size_t level) override;
+        void storeOnLevelVelocity(std::ofstream& os, size_t level);
+        void storeOnLevelOmega(std::ofstream& os, size_t level);
+        void storeOnLevelRefPoints(std::ofstream& os, size_t level);
+        void storeOnLevelRefCurves(std::ofstream& os, size_t level);
+        void storeOnLevelRefSurfaces(std::ofstream& os, size_t level);
         void storeOnTimeSeries(std::ofstream& os) override;
-        FColDsptr getVelocity3D(int i);
-        FColDsptr getOmega3D(int i);
+        FColDsptr getVelocity3D(size_t i);
+        FColDsptr getOmega3D(size_t i);
 
         FColDsptr velocity3D = std::make_shared<FullColumn<double>>(3);
         FColDsptr omega3D = std::make_shared<FullColumn<double>>(3);
